requirements.txt
pygame==2.6.1
pillow==11.0.0
